feat(matchengine): position-aware power formula, lineup sync, and match-result Moral
- Diagnosed and fixed the core disconnect: Club.LineupJson/Formation
  (ClubManagement) never reached MatchEngine as a domain event, so the
  power calculation could never be position/lineup-aware. Added
  ILineupUpdatedEvent, published from Club.UpdateLineup/UpdateFormation,
  consumed by a new LineupUpdatedEventConsumer in MatchEngine.
- Redesigned ClubPowerRating: replaced the flat TotalOverall sum with a
  real roster snapshot (RosterPlayerSnapshot: PlayerId/Overall/Position)
  and a lineup snapshot (LineupSlotAssignment). This also fixes the
  long-standing "can't remove player without knowing Overall" limitation
  in PlayerRemovedFromRosterCommandConsumer.
- New IClubPowerCalculator implements the deepened formula: BazGuc
  (Starting XI, Overall x position-compatibility multiplier), YildizBonusu
  (star player effect), DerinlikBonusu (bench depth), MoralModifiye.
  FormationCatalog + PositionCompatibility encode the 4-tier multiplier
  table (exact match / same family / adjacent family / distant family).
  Missing/incomplete lineups auto-fill from the best remaining roster
  players instead of being penalized as broken.
- New persistent Moral field ([-5,+5], win/loss/draw driven via
  ApplyMatchResult) on ClubPowerRating, separate from the existing
  one-shot weekly MoraleBonus; both feed into MoralModifiye.
  MatchSimulator now takes double team power instead of int, unchanged
  home-advantage math on top.
- PlayerAddedToRosterEvent/IPlayerAddedToRosterEvent now carry Position.
- New EF Core migration RedesignClubPowerRating.

Verified with unit tests (ClubPowerCalculatorTests: position-multiplier
table, deterministic + 30-match statistical broken-vs-correct-lineup
series, Moral clamp) and a real E2E script
(tests/run_e2e_test_powerrating.ps1) against live Draft->Saga->
ClubManagement->RabbitMQ->MatchEngine->Postgres infra, DB-verified via
psql: same squad, same opponent, correct lineup 7W-3D-4L/24pts vs broken
lineup (GK/CB played out of position) 3W-4D-7L/13pts over 14 matches
each; Moral repeatedly reaching and holding +5 across real matches.

+++ src/Services/MatchEngine/Domain/Services/IMatchSimulator.cs
@@ -4,5 +4,5 @@ namespace ClubCraft.MatchEngine.Domain.Services;
 
 public interface IMatchSimulator
 {
-    void Simulate(Match match, int homePower, int awayPower);
+    void Simulate(Match match, double homePower, double awayPower);
 }
